Conflicts:
	plugins/org.jkiss.dbeaver.core/OSGI-INF/l10n/bundle_fr.properties
	plugins/org.jkiss.dbeaver.data.transfer/META-INF/MANIFEST.MF
	plugins/org.jkiss.dbeaver.data.transfer/OSGI-INF/l10n/bundle.properties
	plugins/org.jkiss.dbeaver.data.transfer/src/org/jkiss/dbeaver/tools/transfer/stream/StreamProducerSettings.java
	plugins/org.jkiss.dbeaver.data.transfer/src/org/jkiss/dbeaver/tools/transfer/stream/StreamTransferConsumer.java
	plugins/org.jkiss.dbeaver.data.transfer/src/org/jkiss/dbeaver/tools/transfer/stream/exporter/DataExporterSQL.java
	plugins/org.jkiss.dbeaver.data.transfer/src/org/jkiss/dbeaver/tools/transfer/stream/exporter/DataExporterTXT.java
	plugins/org.jkiss.dbeaver.ext.generic.ui/src/org/jkiss/dbeaver/ext/generic/views/GenericConnectionPage.java
	plugins/org.jkiss.dbeaver.ext.oracle/src/org/jkiss/dbeaver/ext/oracle/data/OracleTemporalAccessorValueHandler.java
	plugins/org.jkiss.dbeaver.ext.oracle/src/org/jkiss/dbeaver/ext/oracle/model/OracleSQLDialect.java
	plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/DBValueFormatting.java
	plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/exec/DBExecUtils.java
	plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/navigator/DBNDatabaseNode.java
	plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/sql/SQLUtils.java
	plugins/org.jkiss.dbeaver.registry/src/org/jkiss/dbeaver/registry/driver/DriverDescriptor.java
	plugins/org.jkiss.dbeaver.registry/src/org/jkiss/dbeaver/registry/driver/DriverLibraryMavenArtifact.java
	plugins/org.jkiss.dbeaver.ui.app.standalone/splash.bmp
	plugins/org.jkiss.dbeaver.ui.editors.connection/src/org/jkiss/dbeaver/ui/dialogs/connection/ConnectionPageAbstract.java
	plugins/org.jkiss.dbeaver.ui.editors.data/src/org/jkiss/dbeaver/ui/controls/resultset/ResultSetPersister.java
	plugins/org.jkiss.dbeaver.ui.editors.data/src/org/jkiss/dbeaver/ui/controls/resultset/internal/ResultSetMessages_fr.properties
	plugins/org.jkiss.dbeaver.ui.editors.data/src/org/jkiss/dbeaver/ui/controls/resultset/spreadsheet/SpreadsheetPresentation.java
	plugins/org.jkiss.dbeaver.ui.editors.sql/src/org/jkiss/dbeaver/ui/editors/sql/SQLEditor.java
	plugins/org.jkiss.dbeaver.ui.editors.sql/src/org/jkiss/dbeaver/ui/editors/sql/execute/SQLQueryJob.java
	product/debug/DBeaver.product.launch

+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/impl/app/DefaultCertificateStorage.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2020 DBeaver Corp and others
+ * Copyright (C) 2010-2021 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 package org.jkiss.dbeaver.model.impl.app;
 
 import org.jkiss.code.NotNull;
+import org.jkiss.code.Nullable;
 import org.jkiss.dbeaver.DBException;
 import org.jkiss.dbeaver.Log;
 import org.jkiss.dbeaver.model.DBPDataSourceContainer;
@@ -28,8 +29,7 @@ import java.security.*;
 import java.security.cert.Certificate;
 import java.security.cert.CertificateFactory;
 import java.security.spec.PKCS8EncodedKeySpec;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 /**
  * DefaultCertificateStorage
@@ -37,7 +37,7 @@ import java.util.List;
 public class DefaultCertificateStorage implements DBACertificateStorage {
 
     private static final Log log = Log.getLog(DefaultCertificateStorage.class);
-    public static final char[] DEFAULT_PASSWORD = "".toCharArray();
+    private static final char[] DEFAULT_PASSWORD = "".toCharArray();
     public static final String JKS_EXTENSION = ".jks";
 
     public static final String CA_CERT_ALIAS = "ca-cert";
@@ -45,9 +45,11 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
     public static final String KEY_CERT_ALIAS = "key-cert";
 
     private final File localPath;
+    private final Map<String, UserDefinedKeystore> userDefinedKeystores;
 
     public DefaultCertificateStorage(File localPath) {
         this.localPath = localPath;
+        this.userDefinedKeystores = new HashMap<>();
         if (localPath.exists()) {
             // Cleanup old keystores
             final File[] ksFiles = localPath.listFiles();
@@ -70,7 +72,7 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
             KeyStore ks = KeyStore.getInstance(KeyStore.getDefaultType());
             if (ksFile.exists()) {
                 try (InputStream is = new FileInputStream(ksFile)) {
-                    ks.load(is, DEFAULT_PASSWORD);
+                    ks.load(is, getKeyStorePassword(container, certType));
                 }
             } else {
                 ks.load(null, DEFAULT_PASSWORD);
@@ -112,6 +114,9 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
 
     @Override
     public void addCertificate(@NotNull DBPDataSourceContainer dataSource, @NotNull String certType, byte[] caCertData, byte[] clientCertData, byte[] keyData) throws DBException {
+        if (userDefinedKeystores.containsKey(getKeyStoreName(dataSource, certType))) {
+            throw new DBException("Adding new certificates would override user-specified keystore");
+        }
         final KeyStore keyStore = getKeyStore(dataSource, certType);
         try {
             CertificateFactory cf = CertificateFactory.getInstance("X.509");
@@ -128,7 +133,7 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
             }
             if (keyData != null) {
                 PrivateKey privateKey = loadPrivateKeyFromPEM(keyData);
-               keyStore.setKeyEntry(KEY_CERT_ALIAS, privateKey, DEFAULT_PASSWORD, certChain.toArray(new Certificate[certChain.size()]));
+                keyStore.setKeyEntry(KEY_CERT_ALIAS, privateKey, DEFAULT_PASSWORD, certChain.toArray(new Certificate[certChain.size()]));
             }
 
             saveKeyStore(dataSource, certType, keyStore);
@@ -139,7 +144,18 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
     }
 
     @Override
+    public void addCertificate(@NotNull DBPDataSourceContainer dataSource, @NotNull String certType, @NotNull String keyStorePath, @NotNull char[] keyStorePassword) throws DBException {
+        userDefinedKeystores.put(
+            getKeyStoreName(dataSource, certType),
+            new UserDefinedKeystore(new File(keyStorePath), keyStorePassword)
+        );
+    }
+
+    @Override
     public void addSelfSignedCertificate(@NotNull DBPDataSourceContainer dataSource, @NotNull String certType, @NotNull String certDN) throws DBException {
+        if (userDefinedKeystores.containsKey(getKeyStoreName(dataSource, certType))) {
+            throw new DBException("Adding new certificates would override user-specified keystore");
+        }
         final KeyStore keyStore = getKeyStore(dataSource, certType);
         try {
             List<Certificate> certChain = new ArrayList<>();
@@ -156,12 +172,16 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
 
             saveKeyStore(dataSource, certType, keyStore);
         } catch (Throwable e) {
-            throw new DBException("Error adding self signed certificate to keystore", e);
+            throw new DBException("Error adding self-signed certificate to keystore", e);
         }
     }
 
     @Override
     public void deleteCertificate(@NotNull DBPDataSourceContainer dataSource, @NotNull String certType) throws DBException {
+        if (userDefinedKeystores.remove(getKeyStoreName(dataSource, certType)) != null) {
+            // We don't want to erase anything from user-defined keystore
+            return;
+        }
         final KeyStore keyStore = getKeyStore(dataSource, certType);
         try {
             keyStore.deleteEntry(CA_CERT_ALIAS);
@@ -175,7 +195,23 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
 
     @Override
     public File getKeyStorePath(DBPDataSourceContainer dataSource, String certType) {
-        return new File(localPath, dataSource.getId() + "-" + certType + JKS_EXTENSION);
+        final UserDefinedKeystore userDefinedKeystore = getUserDefinedKeystore(dataSource, certType);
+        if (userDefinedKeystore != null) {
+            return userDefinedKeystore.file;
+        } else {
+            return new File(localPath, getKeyStoreName(dataSource, certType) + JKS_EXTENSION);
+        }
+    }
+
+    @NotNull
+    @Override
+    public char[] getKeyStorePassword(@NotNull DBPDataSourceContainer dataSource, @NotNull String certType) {
+        final UserDefinedKeystore userDefinedKeystore = getUserDefinedKeystore(dataSource, certType);
+        if (userDefinedKeystore != null) {
+            return userDefinedKeystore.password;
+        } else {
+            return DEFAULT_PASSWORD;
+        }
     }
 
     @Override
@@ -219,4 +255,23 @@ public class DefaultCertificateStorage implements DBACertificateStorage {
         }
     }
 
+    @Nullable
+    private UserDefinedKeystore getUserDefinedKeystore(@NotNull DBPDataSourceContainer dataSource, @NotNull String certType) {
+        return userDefinedKeystores.get(getKeyStoreName(dataSource, certType));
+    }
+
+    @NotNull
+    private String getKeyStoreName(@NotNull DBPDataSourceContainer dataSource, @NotNull String certType) {
+        return dataSource.getId() + '-' + certType;
+    }
+
+    private static class UserDefinedKeystore {
+        private final File file;
+        private final char[] password;
+
+        public UserDefinedKeystore(@NotNull File file, @NotNull char[] password) {
+            this.file = file;
+            this.password = password;
+        }
+    }
 }
